Go Forward and Back in Quotes
With lots of quotes its hard to find the one you want. This introduces a
quotes presenter which wraps up the logic of presenting the quote and
the next and previous quotes to the view.

It selects the quotes by limit queries and wraps around so clicking
previous on the first one goes to the last one and vice versa.